change rgb to big endian to avoid geoserver png compression on aplpha=0

+++ server/guppy2/endpoint_utils.py
@@ -71,4 +71,4 @@ def _decode(data):
     """
     Utility to decode RGB encoded data
     """
-    return np.frombuffer(data.reshape(4, -1).transpose().tobytes(), dtype='>f4').reshape((data[0].shape))
+    return np.frombuffer(data.reshape(4, -1).transpose().tobytes(), dtype='<f4').reshape((data[0].shape))
